bug(fix-bug): to convert response to json
- changing response to json
- renaming swaggersetup to ah-92explorers
- Change the url

[Maintains #166954780]
