+++ controllers/welcomeController.js
@@ -4,12 +4,9 @@ dotenv.config();
 
 class WelcomeController {
   static welcome(req, res) {
-    res
-      .status(200)
-      .send(
-        "Welcome to Authors Haven - A Social platform for the creative at heart."
-      );
-    res.status(400).send("Server Error");
+    res.status(200).json({
+      msg: "Welcome to Authors Haven API, a #92Explorers Product."
+    });
   }
 }
